# Remove unnecessary initialization of create_id in JSON.parse()